refactor : Ladder에서 people 변수 추출 및 people 객체 생성
