+++ src/main/java/nextstep/ladder/model/Ladder.java
@@ -1,28 +1,19 @@
 package nextstep.ladder.model;
 
 import java.util.Arrays;
+import java.util.List;
 
 public class Ladder {
-    private String[] people;
     private int height;
     private Lines lines;
 
-    public Ladder(String people, int height){
-        this.people = Arrays.stream(people.split(","))
-                            .map(String::trim)
-                            .toArray(String[]::new);
+    public Ladder(int countOfPerson, int height){
         this.height = height;
-        lines = new Lines(this.people.length, this.height);
+        lines = new Lines(countOfPerson, this.height);
 
     }
-
-    public String[] getPeople(){
-        return people;
-    }
-
-
-    public Lines getLines(){
-        return lines;
+    public List<Line> getLines(){
+        return lines.getLines();
     }
